--- libs/ledger/src/ledger/lane_controller.cpp
@@ -1,51 +1,0 @@
-//------------------------------------------------------------------------------
-//
-//   Copyright 2018-2019 Fetch.AI Limited
-//
-//   Licensed under the Apache License, Version 2.0 (the "License");
-//   you may not use this file except in compliance with the License.
-//   You may obtain a copy of the License at
-//
-//       http://www.apache.org/licenses/LICENSE-2.0
-//
-//   Unless required by applicable law or agreed to in writing, software
-//   distributed under the License is distributed on an "AS IS" BASIS,
-//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//   See the License for the specific language governing permissions and
-//   limitations under the License.
-//
-//------------------------------------------------------------------------------
-
-#include "ledger/storage_unit/lane_controller.hpp"
-
-namespace fetch {
-namespace ledger {
-
-void LaneController::WorkCycle()
-{
-  UriSet remove;
-  UriSet create;
-
-  GeneratePeerDeltas(create, remove);
-
-  FETCH_LOG_WARN(LOGGING_NAME, "WorkCycle:create:", create.size());
-  FETCH_LOG_WARN(LOGGING_NAME, "WorkCycle:remove:", remove.size());
-
-  for (auto &uri : create)
-  {
-    FETCH_LOG_WARN(LOGGING_NAME, "WorkCycle:creating:", uri.ToString());
-    muddle_->AddPeer(Uri(uri.ToString()));
-  }
-
-  for (auto &uri : create)
-  {
-    Address target_address;
-    if (muddle_->UriToDirectAddress(uri, target_address))
-    {
-      peer_connections_[uri] = target_address;
-    }
-  }
-}
-
-}  // namespace ledger
-}  // namespace fetch
